pyproject.toml
[project]
name = "open-alita"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=0.8.0",
    "duckduckgo-search>=8.0.0",
    "fastapi>=0.110.0",
    "firecrawl-py>=0.0.16",
    "flask>=3.1.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.3",
    "json-repair>=0.7.0",
    "langchain-community>=0.3.19",
    "langchain-core>=0.1.0",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai>=0.3.8",
    "langgraph>=0.4.3",
    "litellm>=1.63.11",
    "loguru>=0.6.0",
    "markdownify>=1.1.0",
    "mcp>=1.6.0",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "pydantic>=2.0.0,<3.0.0",
    "pillow>=10.0.0",
    "playwright>=1.40.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.0.1",
    "readabilipy>=0.3.0",
    "requests>=2.32.3",
    "socksio>=1.0.0",
    "sse-starlette>=1.6.5",
    "uvicorn>=0.27.1",
    "browser-use==0.4.2",
    "openpyxl>=3.1.5",
    "none>=0.1.1",
    "huggingface-hub>=0.33.0",
    "selenium>=4.33.0",
    "pytesseract>=0.3.13",
    "pypdf2>=3.0.1",
]
